+++ CHANGELOG.md
@@ -376,3 +376,4 @@ with new `PluginSerializer`
 * `PostsUsedTable` now is deprecated
 * `FixPost` will remove transaction only in case of availability of it
 * `BasePlugin` now work with internal lambdas and coroutine jobs
+* `CommonKnownPostsTransactions` was created